+++ packages/cli/src/commands/bisect/index.ts
@@ -16,7 +16,7 @@ import Client from '../../util/client';
 import { getPkgName } from '../../util/pkg-name';
 import { Output } from '../../util/output';
 import { Deployment, PaginationOptions } from '../../types';
-import { normalizeURL } from '../../util/commands/bisect';
+import { normalizeURL } from '../../util/bisect/normalize-url';
 
 interface DeploymentV6
   extends Pick<
